agents/comparative_analysis_agent/requirements.txt
fastapi==0.104.1
uvicorn==0.24.0
sqlalchemy==2.0.23
psycopg2-binary==2.9.9
python-dotenv==1.0.0
pydantic==2.5.0
aiohttp==3.9.1
requests==2.31.0
numpy==1.25.2
pandas==2.1.4
scipy==1.11.4

# Multi-Tool Integration Dependencies
# LangChain Integration
langchain==0.1.0
langchain-openai==0.0.2
langchain-community==0.0.10
openai==1.3.7

# Computer Use Integration
computer-use==0.1.0

# LlamaIndex Integration
llama-index==0.9.8
llama-index-embeddings-openai==0.1.6
llama-index-llms-openai==0.1.12
llama-index-vector-stores-chroma==0.1.4

# Haystack Integration
farm-haystack==1.21.2
sentence-transformers==2.2.2

# AutoGen Integration
pyautogen==0.2.0

# Additional Dependencies
chromadb==0.4.18
tiktoken==0.5.1
transformers==4.35.2
torch==2.1.1
scikit-learn==1.3.2 